refactor(auth): kill adapter layer, inline types into factory
Delete types.ts, local-storage-adapter.svelte.ts from both apps.
Factory now takes storagePrefix instead of adapter instances and
creates createPersistedState internally. AuthUser, AuthPhase, and
AuthError inlined into the factory—no separate types file.

Per-app wiring shrinks from 61 lines to 28 lines. 4 files → 2 files
per app. No interface, no adapter, no indirection.
