+++ apps/honeycrisp/package.json
@@ -24,8 +24,7 @@
 		"svelte-sonner": "catalog:",
 		"tailwindcss": "catalog:",
 		"typescript": "catalog:",
-		"vite": "catalog:",
-		"@standard-schema/spec": "catalog:"
+		"vite": "catalog:"
 	},
 	"dependencies": {
 		"@epicenter/svelte": "workspace:*",
