+++ apps/opensidian/package.json
@@ -19,7 +19,6 @@
 		"wellcrafted": "catalog:"
 	},
 	"devDependencies": {
-		"@standard-schema/spec": "catalog:",
 		"@codemirror/commands": "^6.10.3",
 		"@codemirror/lang-markdown": "^6.5.0",
 		"@codemirror/language": "^6.12.2",
